requirements.txt
langchain
streamlit
langchain_openai
python-dotenv
